refactor: move helper function to main function
The `button-create` file had a helper function `button-create-container`.
Merged them into one file for better file organization within the
components folder. This change helps in tracking related functions more
effectively.
